fixes.
Previous massive parallel modifications :
Corrected the fetching errors with retrial (up to 10 times).
Using Runtime.availableProcessors() * 10 as parallelism level.
Less checks on orders fetch returning null, those are made in the fetch.

Reduced the debug output(actually moved usual debug level to trace).

Corrected the BOSO value of LP offers :
This is the required items'BO value for ONE item (because I only place
ONE BO a.1 isk more than the highest one).
Similarly the produced items'SO for one item.

Corrected use of broker and sale tax for SOBO and BOSO :
The default sale has been set to 1%, the default broker to 2%.
Using double value in the formula to avoid rounding the integer.