big modifications in tree and classes
1. change name from ESI to JCESI
2. added a module JCESI-interfaces that provides the interfaces
previously generated by the compiler. The compiler thus becomes easier
to understand.
3. added a ObsX Holder, for X=list;map;Object that replace the
observableX returns by the cache. Those classes have waitdata() and
follows(listener) as well as copy() that allow to work with them without
using synchronized (it is embedded in the methods)

However a few synch bug make it unusable ATM